feat: Enhance Kotlin Multiplatform SDK with new features and improvements
- Updated Kotlin version to 2.1.0 for better performance and compatibility.
- Introduced a new in-memory database implementation for efficient data handling.
- Added platform-specific time utilities for Android and JVM to standardize time management.
- Implemented a comprehensive event bus system for better event handling across components.
- Enhanced the Whisper STT service with improved error handling and logging.
- Cleaned up unused code and configurations, improving maintainability and clarity.
- Updated documentation to reflect new features and architectural changes, ensuring clarity for future development.

+++ sdk/runanywhere-kotlin/src/commonMain/kotlin/com/runanywhere/sdk/models/ModelLoadingService.kt
@@ -109,7 +109,7 @@ class ModelLoadingService(
 
         // Emit progress events
         for (progress in 1..10) {
-            EventBus.publish(SDKModelEvent.DownloadProgress(modelInfo.id, progress / 10f))
+            EventBus.publish(SDKModelEvent.DownloadProgress(modelInfo.id, (progress / 10f).toDouble()))
             kotlinx.coroutines.delay(100) // Simulate download time
         }
     }
